Extract Line utils
