+++ src/utils/MapConverter.ts
@@ -1,10 +1,13 @@
+import type { Point } from './lineUtils'
+import type { Position } from 'geojson';
+
 export class MapConverter {
   constructor(
     private mapWidth: number,
     private mapHeight: number
   ) {}
 
-  public svgCoordsFromGeoCoords(geoCoords: number[]) {
+  public svgCoordsFromGeoCoords(geoCoords: Point | Position): Point {
     const [lon, lat] = geoCoords;
     const x = (lon + 180) * (this.mapWidth / 360);
     const y = ((lat * -1) + 90) * (this.mapHeight / 180);
