+++ src/utils/connectDots.ts
@@ -1,5 +1,4 @@
-type Point = [number, number]
-type Line = [Point, Point]
+import { Point, Line, areLinesIntersecting, getLineLength } from './lineUtils'
 
 export function connectDots(points: Point[]) {
   const lines = connectAllToAll(points)
@@ -9,7 +8,7 @@ export function connectDots(points: Point[]) {
       if (linesToRemove.includes(j) || linesToRemove.includes(i)) {
         continue
       }
-      if (areIntersecting(lines[i], lines[j])) {
+      if (areLinesIntersecting(lines[i], lines[j])) {
         if (getLineLength(lines[i]) > getLineLength(lines[j])) {
           linesToRemove.push(i)
         } else {
@@ -29,25 +28,4 @@ function connectAllToAll(points: Point[]) {
     }
   }
   return lines
-}
-
-function getLineLength(line: Line) {
-  const [p1, p2] = line
-  const x = p1[0] - p2[0]
-  const y = p1[1] - p2[1]
-  return Math.sqrt(x * x + y * y)
-}
-
-function areIntersecting(line1: Line, line2: Line): boolean {
-  const [[a, b], [c, d]] = line1;
-  const [[p, q], [r, s]] = line2;
-  
-  const det = (c - a) * (s - q) - (r - p) * (d - b);
-  if (det === 0) {
-    return false;
-  } else {
-    const lambda = ((s - q) * (r - a) + (p - r) * (s - b)) / det;
-    const gamma = ((b - d) * (r - a) + (c - a) * (s - b)) / det;
-    return (0 < lambda && lambda < 1) && (0 < gamma && gamma < 1);
-  }
 }
